PLX-1008 added pino logger, refactored core logger lib
