+++ package.json
@@ -48,5 +48,8 @@
     "supertest": "^6.3.3",
     "ts-jest": "^29.1.1",
     "typescript": "^5.1.6"
+  },
+  "dependencies": {
+    "pino": "^8.14.2"
   }
 }
