+++ src/app/server/package.json
@@ -26,6 +26,7 @@
     "@koa/router": "^12.0.0",
     "axios": "1.1.3",
     "dotenv": "^16.0.3",
-    "koa": "^2.13.4"
+    "koa": "^2.13.4",
+    "koa-pino-logger": "^4.0.0"
   }
 }
