--- src/app/server/src/gateways/httpGateway/index.js
@@ -1,138 +1,0 @@
-"use strict";
-var __assign = (this && this.__assign) || function () {
-    __assign = Object.assign || function(t) {
-        for (var s, i = 1, n = arguments.length; i < n; i++) {
-            s = arguments[i];
-            for (var p in s) if (Object.prototype.hasOwnProperty.call(s, p))
-                t[p] = s[p];
-        }
-        return t;
-    };
-    return __assign.apply(this, arguments);
-};
-var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, generator) {
-    function adopt(value) { return value instanceof P ? value : new P(function (resolve) { resolve(value); }); }
-    return new (P || (P = Promise))(function (resolve, reject) {
-        function fulfilled(value) { try { step(generator.next(value)); } catch (e) { reject(e); } }
-        function rejected(value) { try { step(generator["throw"](value)); } catch (e) { reject(e); } }
-        function step(result) { result.done ? resolve(result.value) : adopt(result.value).then(fulfilled, rejected); }
-        step((generator = generator.apply(thisArg, _arguments || [])).next());
-    });
-};
-var __generator = (this && this.__generator) || function (thisArg, body) {
-    var _ = { label: 0, sent: function() { if (t[0] & 1) throw t[1]; return t[1]; }, trys: [], ops: [] }, f, y, t, g;
-    return g = { next: verb(0), "throw": verb(1), "return": verb(2) }, typeof Symbol === "function" && (g[Symbol.iterator] = function() { return this; }), g;
-    function verb(n) { return function (v) { return step([n, v]); }; }
-    function step(op) {
-        if (f) throw new TypeError("Generator is already executing.");
-        while (g && (g = 0, op[0] && (_ = 0)), _) try {
-            if (f = 1, y && (t = op[0] & 2 ? y["return"] : op[0] ? y["throw"] || ((t = y["return"]) && t.call(y), 0) : y.next) && !(t = t.call(y, op[1])).done) return t;
-            if (y = 0, t) op = [op[0] & 2, t.value];
-            switch (op[0]) {
-                case 0: case 1: t = op; break;
-                case 4: _.label++; return { value: op[1], done: false };
-                case 5: _.label++; y = op[1]; op = [0]; continue;
-                case 7: op = _.ops.pop(); _.trys.pop(); continue;
-                default:
-                    if (!(t = _.trys, t = t.length > 0 && t[t.length - 1]) && (op[0] === 6 || op[0] === 2)) { _ = 0; continue; }
-                    if (op[0] === 3 && (!t || (op[1] > t[0] && op[1] < t[3]))) { _.label = op[1]; break; }
-                    if (op[0] === 6 && _.label < t[1]) { _.label = t[1]; t = op; break; }
-                    if (t && _.label < t[2]) { _.label = t[2]; _.ops.push(op); break; }
-                    if (t[2]) _.ops.pop();
-                    _.trys.pop(); continue;
-            }
-            op = body.call(thisArg, _);
-        } catch (e) { op = [6, e]; y = 0; } finally { f = t = 0; }
-        if (op[0] & 5) throw op[1]; return { value: op[0] ? op[1] : void 0, done: true };
-    }
-};
-var __importDefault = (this && this.__importDefault) || function (mod) {
-    return (mod && mod.__esModule) ? mod : { "default": mod };
-};
-Object.defineProperty(exports, "__esModule", { value: true });
-exports.HttpGateway = void 0;
-var axios_1 = __importDefault(require("axios"));
-var logger_1 = require("../../../../../core/logger");
-var log = (0, logger_1.createLogger)('httpGateway');
-var HttpGateway = /** @class */ (function () {
-    function HttpGateway(requestConfig) {
-        this.requestConfig = requestConfig;
-    }
-    HttpGateway.prototype.request = function (requestLocalConfig) {
-        return __awaiter(this, void 0, void 0, function () {
-            var url, config, axiosResponse, error_1;
-            return __generator(this, function (_a) {
-                switch (_a.label) {
-                    case 0:
-                        url = "".concat(this.requestConfig.baseUrl).concat(requestLocalConfig.endpoint);
-                        config = {
-                            headers: this.buildRequestHeaders(requestLocalConfig.headers),
-                            params: __assign(__assign({}, this.requestConfig.queryParams), requestLocalConfig.queryParams),
-                            timeout: this.requestConfig.timeout
-                        };
-                        log.debug({ url: url, config: config }, 'HttpGateway request');
-                        _a.label = 1;
-                    case 1:
-                        _a.trys.push([1, 10, , 11]);
-                        if (!(requestLocalConfig.method === 'GET')) return [3 /*break*/, 3];
-                        return [4 /*yield*/, axios_1.default.get(url, config)];
-                    case 2:
-                        axiosResponse = _a.sent();
-                        return [3 /*break*/, 9];
-                    case 3:
-                        if (!(requestLocalConfig.method === 'POST')) return [3 /*break*/, 5];
-                        return [4 /*yield*/, axios_1.default.post(url, requestLocalConfig.body, config)];
-                    case 4:
-                        axiosResponse = _a.sent();
-                        return [3 /*break*/, 9];
-                    case 5:
-                        if (!(requestLocalConfig.method === 'PUT')) return [3 /*break*/, 7];
-                        return [4 /*yield*/, axios_1.default.put(url, requestLocalConfig.body, config)];
-                    case 6:
-                        axiosResponse = _a.sent();
-                        return [3 /*break*/, 9];
-                    case 7:
-                        if (!(requestLocalConfig.method === 'DELETE')) return [3 /*break*/, 9];
-                        return [4 /*yield*/, axios_1.default.delete(url, config)];
-                    case 8:
-                        axiosResponse = _a.sent();
-                        _a.label = 9;
-                    case 9: return [2 /*return*/, {
-                            status: axiosResponse.status,
-                            statusText: axiosResponse.statusText,
-                            data: axiosResponse.data,
-                            headers: axiosResponse.headers
-                        }];
-                    case 10:
-                        error_1 = _a.sent();
-                        return [2 /*return*/, this.handleError(error_1)];
-                    case 11: return [2 /*return*/];
-                }
-            });
-        });
-    };
-    HttpGateway.prototype.buildRequestHeaders = function (requestLocalConfigHeaders) {
-        return __assign(__assign({ 'Content-Type': 'application/json' }, this.requestConfig.headers), requestLocalConfigHeaders);
-    };
-    HttpGateway.prototype.kjhdaf = function () {
-        return undefined;
-    };
-    HttpGateway.prototype.handleError = function (error) {
-        if (!axios_1.default.isAxiosError(error) || error.response == null) {
-            var gatewayError = new Error('Unexpected error' + ((error === null || error === void 0 ? void 0 : error.message) === null ? '' : ": ".concat(error.message)));
-            gatewayError.name = 'Gateway Error';
-            log.error({ gatewayError: gatewayError, error: error });
-            throw gatewayError;
-        }
-        log.warn({ errorResponse: error.response }, 'HttpGateway request error');
-        return {
-            status: error.response.status,
-            statusText: error.response.statusText,
-            data: error.response.data,
-            headers: error.response.headers
-        };
-    };
-    return HttpGateway;
-}());
-exports.HttpGateway = HttpGateway;
-//# sourceMappingURL=index.js.map
