+++ src/core/logger/logger.interface.ts
@@ -1,0 +1,8 @@
+type loggerFunction = (params: Record<string, any> | string, message?: string | undefined) => void
+
+export interface Logger {
+  debug: loggerFunction
+  info: loggerFunction
+  warn: loggerFunction
+  error: loggerFunction
+}
